% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{Rmse}
\alias{Rmse}
\title{calcula el RMSE o NRMSE}
\usage{
Rmse(imp, mis, true, norm = FALSE)
}
\arguments{
\item{imp}{la data imputada}

\item{mis}{la data con missings}

\item{true}{la verdadera data matriz}

\item{norm}{logical, si eS TRUE ENTONCES EL RMSE volverá}
}
\value{
the RMSE or NRMSE
}
\description{
This function calculate imputation error given the imputed data, the missing
data and the true data
}
\examples{
data(parkinson)
# introduce 10\% random missing values into the parkinson data
missdata <- SimIm(parkinson, 0.1)

# impute the missing values by LASSO
\donttest{impdata <- impute(missdata, lmFun = "lassoR")

# calculate the normalised RMSE for the imputation
Rmse(impdata$imp, missdata, parkinson, norm = TRUE)
}
}
\seealso{
\code{\link{impute}} por la funcion de imputacion,
\code{\link{mr}} pr el error de clasificar mal el error metrico
}
