% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{SimIm}
\alias{SimIm}
\title{Introducir de una manera aleatoria valores missings}
\usage{
SimIm(data, p = 0.1)
}
\arguments{
\item{data}{una matriz da data para simular}

\item{p}{el porcentaje de missing que se introducira a la data, esto será en el rango de 0 a 1}
}
\value{
el mismo tamaño de la matriz con los valores missing simulados
}
\description{
Introducir de una manera aleatoria valores missings
}
\examples{
# Crea una data sin los missing como ejemplo
simdata <- matrix(rnorm(100), 10, 10)

# Introduce los missing con un p de 0.15
missingdata <- SimIm(simdata, p = 0.15)

# cuenta el numero de missing values
sum(is.na(missingdata))

#------------------

# No hay datos faltantes en la data original de parkinson
data(parkinson)

# Introduzcamos ahora datos faltantes
missdata <- SimIm(parkinson, 0.1)

# Cuenta ahora el numero de missings en la nueva data
sum(is.na(missdata))
}
\keyword{imputation}
\keyword{simulation,}
